refactor: useDeviceOrientation hook
