+++ src/components/NovelList.tsx
@@ -9,7 +9,7 @@ import {
 } from 'react-native';
 import { NovelItem } from '@plugins/types';
 import { LibraryNovelInfo, NovelInfo } from '../database/types';
-import { useDeviceOrientation } from '@hooks/useDeviceOrientation';
+import { useDeviceOrientation } from '@hooks';
 
 export type NovelListRenderItem = ListRenderItem<
   LibraryNovelInfo | NovelInfo | NovelItem
