+++ src/hooks/index.ts
@@ -10,6 +10,7 @@ export { useCategorySettings, useLibrarySettings } from './useSettings';
 export { usePreviousRouteName } from './usePreviousRouteName';
 export { useBackHandler } from './common/useBackHandler';
 export { useTextToSpeech, TextToSpeechStatus } from './common/useTextToSpeech';
+export { useDeviceOrientation } from './common/useDeviceOrientation';
 
 // hook types
 export type { UseBooleanReturnType } from './common/useBoolean';
